requirements.txt

apscheduler==3.6.3
cachetools==4.2.2
email-validator>=2.2.0
flask>=3.1.0
matplotlib>=3.10.1
numpy>=2.2.3
openpyxl>=3.1.5
pandas>=2.2.3
phonenumbers>=9.0.0
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
python-telegram-bot==13.7
requests>=2.32.3
tornado==6.1
trafilatura>=2.0.0
twilio>=9.4.6
